fix: :bug: Changes on design and added modal when a route is succesfull added
