+++ src/app/conductores/newConductor/page.jsx
@@ -1,17 +1,28 @@
 "use client"
 import React from "react";
 import { useRouter } from "next/navigation";
+import { Dialog, Transition } from '@headlessui/react'
+import { Fragment, useState } from 'react'
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
+import {faCircleCheck} from '@fortawesome/free-solid-svg-icons'
+
 const NewConductor = () => {
   const router = useRouter();
+  let [isOpen, setIsOpen] = useState(false)
 
   function redirectUser(){
     router.push("http://localhost:3000/conductores/");
   }
 
   function submitNewConductor(){
-    setTimeout(redirectUser, 2000)
+    setIsOpen(true)
+    
   }
-  
+
+  function closeModal() {
+    setIsOpen(false)
+    setTimeout(redirectUser, 500)
+  }
 
   return (
     <article className="flex min-h-screen items-center justify-between bg-[#FFF8F0]">
@@ -67,6 +78,60 @@ const NewConductor = () => {
       <div className="flex flex-col items-center justify-center w-[50%] pb-16">
         <img width={700} src={"/driver.png"} alt="driverHand" />
       </div>
+
+      <Transition appear show={isOpen} as={Fragment}>
+        <Dialog as="div" className="relative rounded-lg z-10" onClose={closeModal}>
+          <Transition.Child
+            as={Fragment}
+            enter="ease-out duration-300"
+            enterFrom="opacity-0"
+            enterTo="opacity-100"
+            leave="ease-in duration-200"
+            leaveFrom="opacity-100"
+            leaveTo="opacity-0"
+          >
+            <div className="fixed inset-0 bg-black bg-opacity-25" />
+          </Transition.Child>
+
+          <div className="flex justify-center items-center fixed inset-0 overflow-y-auto">
+            <div className="flex min-h-full items-center justify-center p-8 text-center">
+              <Transition.Child
+                as={Fragment}
+                enter="ease-out duration-300"
+                enterFrom="opacity-0 scale-95"
+                enterTo="opacity-100 scale-100"
+                leave="ease-in duration-200"
+                leaveFrom="opacity-100 scale-100"
+                leaveTo="opacity-0 scale-95"
+              >
+                <Dialog.Panel className="flex justify-center items-center flex-col w-full max-w-lg transform overflow-hidden rounded-2xl bg-white p-6 text-left align-middle shadow-xl transition-all">
+                  <Dialog.Title
+                    as="h3"
+                    className=" mt-4 text-lg font-medium leading-6 text-gray-700 font-rubik"
+                  >
+                   Conductor agregado con éxito
+                  </Dialog.Title>
+                  <span className="px-4 py-4 inline-flex items-center justify-center text-lg"><FontAwesomeIcon className="text-green-500 h-12 w-12" icon={faCircleCheck}/></span>
+                  <div className="mt-2">
+                    <p className="text-sm text-gray-500 text-center">
+                    No olvides compartir los datos de acceso a tu nuevo conductor
+                    </p>
+                  </div>
+                  <div className="mt-4">
+                    <button
+                      type="button"
+                      className="inline-flex justify-center rounded-md border border-transparent bg-blue-100 px-4 py-2 text-sm font-medium text-blue-900 hover:bg-blue-200 focus:outline-none focus-visible:ring-2 focus-visible:ring-blue-500 focus-visible:ring-offset-2"
+                      onClick={closeModal}
+                    >
+                      Aceptar
+                    </button>
+                  </div>
+                </Dialog.Panel>
+              </Transition.Child>
+            </div>
+          </div>
+        </Dialog>
+      </Transition>
     </article>
   );
 };
